feat(order): 优化订单相关功能和界面
- 修改订单日期选择为仅日期，不包括时间- 更新订单历史记录查看方式- 调整订单详情显示内容
- 修改业务员为录入员
- 优化订单查询界面布局
- 更新订单相关组件样式

+++ order-UI/src/views/dashboard/mixins/order/order_addOrder.js
@@ -27,35 +27,53 @@ export var mixin_order_add = {
 		},
 		// 修改订单的操作
 		handleUpdate(row) {
-			this.$prompt('请输入编辑订单原因', '提示', {
-				confirmButtonText: '确定',
-				cancelButtonText: '取消',
-				type: 'warning'
-			})
-				.then(({ value }) => {
-					addReason({
-						reason: value,
-						tableName: TableName.GOODS_ORDER,
-						tid: row.id,
-						modifyTime: this.modifyTime
-					}).then(() => {
-						// 先暂存订单修改原因
-						sessionStorage.setItem('order-edit-reason', value);
-						this.$message.success('提交成功');
-						this.reset();
-						// 打开弹窗
-						this.openDialog(OrderForm, '修改订单', '1300px', {
-							orderId: row.id,
-							submitInfo: '修改订单'
+			// 获取当前订单日期的时间戳
+			const orderTimestamp = new Date(row.orderDate).getTime();
+			// 设置时间为当天的 00:00:00
+			const today = new Date();
+			today.setHours(0, 0, 0, 0);
+			const todayTimestamp = today.getTime();
+
+			// 如果订单的日期和今天的日期相差超过一天
+			const diffInMs = Math.abs(orderTimestamp - todayTimestamp);
+			const oneDayInMs = 1000 * 60 * 60 * 24; // 一天的毫秒数
+
+			if (diffInMs < oneDayInMs) {
+				// 打开弹窗
+				this.openDialog(OrderForm, '修改订单', '1300px', {
+					orderId: row.id,
+					submitInfo: '修改订单'
+				});
+			} else {
+				this.$prompt('请输入编辑订单原因', '提示', {
+					confirmButtonText: '确定',
+					cancelButtonText: '取消',
+					type: 'warning'
+				})
+					.then(({ value }) => {
+						addReason({
+							reason: value,
+							tableName: TableName.GOODS_ORDER,
+							tid: row.id,
+							modifyTime: this.modifyTime
+						}).then(() => {
+							// 先暂存订单修改原因
+							sessionStorage.setItem('order-edit-reason', value);
+							this.$message.success('提交成功');
+							// 打开弹窗
+							this.openDialog(OrderForm, '修改订单', '1300px', {
+								orderId: row.id,
+								submitInfo: '修改订单'
+							});
+						});
+					})
+					.catch(() => {
+						this.$message({
+							type: 'warning',
+							message: '请先输入编辑原因!'
 						});
 					});
-				})
-				.catch(() => {
-					this.$message({
-						type: 'warning',
-						message: '请先输入编辑原因!'
-					});
-				});
+			}
 		}
 	}
 };
